iOS: kein switch auf GamePhase, CGFloat am Zielbalken
Zwei Reste aus der zweiten CI-Runde. GamePhase ist in Swift eine Klasse,
kein enum — `switch` kann seine Vollstaendigkeit nicht beweisen; der
DYING-Zweig war ohnehin leer und steht jetzt als Kommentar da. Und ein
`goal.fraction` (Kotlin Float) landete ungewandelt an einer CGFloat.

+++ ios/Dottie/Sources/UI/OverlayNodes.swift
@@ -697,7 +697,7 @@ final class GameOverOverlay: SKNode {
             goalLabel.text = L10n.format(
                 "goal_progress", L10n.text(goal.titleKey), Int(goal.current), Int(goal.target)
             )
-            goalBar.fraction = goal.fraction
+            goalBar.fraction = CGFloat(goal.fraction)
         }
         goalLabel.isHidden = goal == nil
         goalBar.isHidden = goal == nil
